% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_package}
\alias{check_package}
\title{checks where a package is within the incoming queues}
\usage{
check_package(package)
}
\arguments{
\item{package}{Package name to check for}
}
\description{
checks where a package is within the incoming queues
}
