% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{incoming}
\alias{incoming}
\title{reports on the contents of CRANs incoming queues}
\usage{
incoming(stage_name = NULL)
}
\arguments{
\item{stage_name}{The name of the queue to check}
}
\description{
reports on the contents of CRANs incoming queues
}
